+++ Module 1 Groovy/src/main/groovy/by/prohor/task1/calculate/expression/Division.groovy
@@ -6,7 +6,7 @@ class Division implements Expression {
     Expression right
 
     @Override
-    double eval() throws Exception {
+    BigDecimal eval() throws Exception {
         return left.eval() / right.eval()
     }
 }
