+++ Module 1 Groovy/src/main/groovy/by/prohor/task1/calculate/expression/atom_value/NumberAtom.groovy
@@ -7,7 +7,7 @@ class NumberAtom extends Atom{
     def value
 
     @Override
-    double eval(){
+    BigDecimal eval(){
         return value
     }
 }
